pyproject.toml
[project]
name = "bahn-de"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "pyyaml>=6.0",
    "jsonschema>=4.10.0"
]
